requirements.txt
joblib==1.4.2
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
